+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support for batch size scaling. ([#691](https://github.com/jina-ai/finetuner/pull/691))
+
 - Add support for learning rate schedulers. ([#679](https://github.com/jina-ai/finetuner/pull/679))
 
 ### Removed