+++ CHANGELOG.md
@@ -12,6 +12,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for batch size scaling. ([#691](https://github.com/jina-ai/finetuner/pull/691))
 
+- Add functions to retrieve evaluation metrics and example results. ([#687](https://github.com/jina-ai/finetuner/pull/687))
+
+### Removed
+
+### Changed
+
+- Remove unit test and integration test from CD. ([#686](https://github.com/jina-ai/finetuner/pull/686))
+
+### Fixed
+
+### Docs
+
+
+## [0.7.2] - 2023-03-02
+
+### Added
+
 - Add support for learning rate schedulers. ([#679](https://github.com/jina-ai/finetuner/pull/679))
 
 ### Removed